--Empty Message--

+++ 2019/prac-2018-r1a/C.cc
@@ -69,7 +69,7 @@ void process() {
     }
   }
   if (buff + base >= P) {
-    printf("%.12lf\n", P);
+    printf("%.12lf\n", (double)P);
   } else {
     printf("%.12lf\n", buff + base);
   }
